File Viewer
Now has a browse button for file viewer. If a radio button is selected before hand, the encryptor will extract raw file data and encrypt it and past it into the output textbox. Future plans on using PrintWriter to create an actual file to put the encrypted data into it.